refactored and modularized some, created some subdirectories in the js directory, added skeleton favicon
