+++ index.html
@@ -4,7 +4,11 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Title</title>
+    
+    <!-- favicon -->
+    <link rel="icon" type="image/png" href="images/favicon.png">
+    
+    <title>Discounts</title>
 
     <!-- fonts -->
     <link href="//fonts.googleapis.com/css?family=Raleway:400,300,600" rel="stylesheet" type="text/css">
