+++ js/discountSelect/option.js
@@ -1,0 +1,7 @@
+const option = (seasonDiscount) => {
+  const opt = document.createElement('option');
+  opt.value = seasonDiscount;
+  opt.innerHTML = seasonDiscount;
+};
+
+module.exports = option;
